backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
PyJWT[crypto]>=2.8.0
python-dotenv>=1.0.0
supabase>=2.0.0
